Update kinet.cpp
Hi, I'm using your code to command two sPDS-480 and I had trouble to communicate with the second power supply and some nodes where not always working on the first one. 

I recorded the packets emitted by "QuickPlay Pro" app and came up with this header for the new protocol version. I works much better for me with this packet header. I do not have access to other ColorKinetics hardware so I can't test it extensively. Since your code is pretty much all that I've been able too find when starting to work on this project, I might be useful for others to add a third protocol. 
Thanks for sharing the original code by the way.